+++ src/App.tsx
@@ -1,4 +1,5 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import LayoutWithNav from '@/pages/layouts/LayoutWithNav';
 import Home from '@/pages/Home';
 import Wiki from '@/pages/Wiki';
@@ -6,6 +7,8 @@ import LayoutWithoutNav from '@/pages/layouts/LayoutWithoutNav';
 import Login from '@/pages/Login';
 import Edit from '@/pages/Edit';
 import SignUp from '@/pages/SignUp';
+
+const queryClient = new QueryClient();
 
 export default function App() {
   const router = createBrowserRouter([
@@ -43,5 +46,9 @@ export default function App() {
     },
   ]);
 
-  return <RouterProvider router={router} />;
+  return (
+    <QueryClientProvider client={queryClient}>
+      <RouterProvider router={router} />
+    </QueryClientProvider>
+  );
 }
